# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Accessible Games Platform API
  version: "1.0.0"
  description: >
    HTTP API for authentication, games catalog, user preferences, follows,
    reviews, and voice intent parsing used by the Team 13 project.
servers:
  - url: http://localhost:5000
    description: Local development
...
paths:
...
  /api/games/search:
    get:
      tags: [Library]
      summary: Search games by text and tags
      parameters:
        - name: q
          in: query
          required: false
          description: Free-text search applied to title and platform.
          schema:
            type: string
          example: puzzle
        - name: tags
          in: query
          required: false
          description: Comma-separated list of tag names; all provided tags must match.
          schema:
            type: string
          example: Vision,Colourblind Mode
      responses:
        "200":
          description: Matching games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameListItem"
        "500":
          description: Search failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
        detail:
          type: string
...
    GameListItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        platform:
          type: string
        releaseDate:
          type: string
          nullable: true
          format: date
        rating:
          type: number
          format: float
        images:
          type: array
          items:
            type: string
            description: Path to thumbnail image
        tags:
          type: array
          items:
            type: string
      example:
        id: 1
        title: Cozy Quest
        platform: PC
        releaseDate: "2024-09-01"
        rating: 4.5
        images:
          - /images/cozy-quest-1.jpg
        tags: [Casual, Large Text, High Contrast]
...